Make AES_KW(P) proofs more self-contained
This patch ensures that the AES_KW(P) proof developments all use files that are
isolated from the proof developments for AES (previously, the former depended
on the latter). In a subsequent commit, we will modify the AES proof
developments to support unbounded AES-GCM, and we want to ensure that changes
to AES do not interfere with the proofs for AES_KW(P).